Organize gadget directory
