+++ polycommit/src/marlin_pc/gadgets/commitment/mod.rs
@@ -14,13 +14,17 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-/// Gadgets for Marlin-KZG10 commitments.
+/// Gadget for an optionally hiding Marlin-KZG10 commitment.
 pub mod commitment;
 pub(crate) use commitment::*;
 
-/// Prepared verifier key gadget for the Marlin-KZG10 polynomial commitment scheme.
-pub mod prepared_verifier_key;
+/// Gadget for a Marlin-KZG10 commitment, with a string label and degree bound.
+pub mod labeled_commitment;
+pub(crate) use labeled_commitment::*;
 
-/// Verifier key gadge for the Marlin-KZG10 polynomial commitment scheme.
-mod verifier_key;
-pub(crate) use verifier_key::*;
+/// Prepared gadget for an optionally hiding Marlin-KZG10 commitment.
+pub mod prepared_commitment;
+pub(crate) use prepared_commitment::*;
+
+/// Prepared gadget for a Marlin-KZG10 commitment, with a string label and degree bound.
+pub mod prepared_labeled_commitment;
